requirements.txt
langchain
langchain-community
transformers
python-dotenv
deep_translator
